feat: add the api endpoint to return latest locked newo
